requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
APScheduler==3.10.4
python-dotenv==1.0.1
twilio==9.2.3
openai>=1.35.0
pytz==2024.1
stripe==7.13.0